+++ Presentation.KuhakuCentral/Controllers/V1/WebScrapingModule/WebScrapingModuleController.cs
@@ -1,7 +1,7 @@
 ﻿using Core.Application.DTOs.General;
 using Core.Application.Features.Scraping.Cuevana.Cuevana3.ch.Commands.GetAllCuevanaMovies;
 using Core.Application.Features.Scraping.PelisPlusLat.Commands.GetPelisPlusLatMovies;
-using KuhakuCentral.Controllers.V1.General;
+using KuhakuCentral.Controllers.General;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Swashbuckle.AspNetCore.Annotations;
@@ -9,7 +9,7 @@ using System.Net.Mime;
 
 namespace KuhakuCentral.Controllers.V1.WebScraping
 {
-    public class ScrapingController : BaseAPI
+    public class WebScrapingModuleController : BaseAPI
     {
 
         [HttpGet("Cuevana3.ch")]
