+++ WebScrapingTesting/WebScrapingTesting.csproj
@@ -16,7 +16,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="HtmlAgilityPack" Version="1.11.59" />
+    <PackageReference Include="HtmlAgilityPack" Version="1.11.61" />
     <PackageReference Include="ScrapySharp" Version="3.0.0" />
   </ItemGroup>
